add 'libft.h' to all .c + correction 'const' in split + some little modification

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: badrien <badrien@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 17:25:39 by badrien           #+#    #+#             */
-/*   Updated: 2019/10/16 14:06:07 by badrien          ###   ########.fr       */
+/*   Updated: 2019/10/22 10:58:11 by badrien          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	int	i;
+	unsigned int	i;
 
 	i = -1;
 	if (dest == 0 && src == 0)
 		return (0);
 	if (dest < src)
-		while ((unsigned int)++i < n)
+		while (++i < n)
 			((unsigned char*)dest)[i] = ((unsigned char*)src)[i];
 	else
 		while (n-- > 0)
